Update sr830.py
- Added write, query, ask functions
- Rewrote other classes to use query/write/ask

Note: if the buffer of the lock-in amplifier is 'corrupt' (i.e. it you submitted a query and cancelled the program before the lock-in was able to return the requested value), the buffer contains information. Any subsequent query will retrieve the first value in the buffer, so that will be the response of the previous query, rather than being the current query. To fix this, issue a .ask() command to retrieve the last value from the buffer, so that subsequent query (write + read) are in sync again.